Fixed the issues with previous push
The HTML5 validator complained in the earlier push, I have given classes to each form input data since they are the same type of data in the add and edit form. I have added specific ID's for each of the edit form data. The ID's for the add form are the same and untouched.