Start,End,Finish Buttons and Time stamps to local storage completed
The start, end, and finish buttons are completed. All buttons are created for each task row but 2 are hidden. When each button is clicked, the style for the other buttons hide or display accordingly. I added 3 new JSON datas. 1. start - which records the start time, 2. end - which records the end time, 3. difference - which records the difference between the end and start. For now, the difference data is in milliseconds and is console logged. The next step is to properly format the data. Some styling also needs to be done for the buttons.