CSS Styling to buttons and forms fixes #74
Fixed the form inputs to be aligned. Made the labels a certain length along with display: inline to achieve this.

Made buttons slightly bigger by adding padding.

Add button was all along aligned with the pageIdentifer section and it aligned based on the text-align center and also has some extra margin-left added to it.