Refactored Code + Fixed issues regarding the date not displaying correctly Fixes #75
Refactored code for the start and end button. Created 2 new functions called startSwitch and endSwitch.

Fixed issues regarding the date not being displayed correctly. There was an issue with the start time being stringified as it was being stored into local storage. Therefore, we had to use a Date.parse on the start time to get the correct format.

Completed meeting notes for this meeting.