Set limit to add with warning popup, Clicking finish button now hides current task Fixes #87, Fixes #86
Hiding display of tasks already finished. Added a new JSON variable called finish. Once finish button is clicked, the finish variable is set to true and when page reloads, will check to see if the task has already been finished. If yes, then it will not load it into the page. If no, then it will load it.

Implemented the limit as to how many items can be in the home page. Set it to 100. Create global variable functions called currentTasks that keeps track of tasks in the home page. In the init function, checks to see if the task is "finished". If it is, then don't count it to finished. If no, then increment "count". Created a MaxLimitPopup along with a button to be clicked to hide the popup again. The form is embedded with the add button and checks to see if the currentTasks is below a certain number. If it is equal to or above, then it will always display the maxpopup and will never display the add form.

Also fixed how the timestamp is presented in log page by using Date.Parse.